fix 2q state tomography after removing qibolabbackend

+++ src/qibocal/auto/transpile.py
@@ -99,7 +99,7 @@ def execute_circuit(
         assert len(sequence.acquisitions) == 1
         result[gate.qubits[0]] = readout[sequence.acquisitions[0][1].id]
 
-    arr = np.stack([result[q] for q in sorted(result)])
+    arr = np.stack([result[q] for q in sorted(result)]).astype(int)
 
     counts = Counter("".join(map(str, col)) for col in arr.T)
